Fix 'Attempt to write to undeclared property' error
Summary:
The latest version of `arc` started throwing an error and quitting when using unit test results via `arc diff` or `arc unit` by saying `Attempt to write to undeclared property parser` in `<disqus>/unittest/DisqusUnitTestEngine.php`.

This patch fixes that by not assigning the temporary test result parser to `$this->parser`. This happens because Arcanist's base Object now has a catch all `__set` method that captures any properties that were not defined explicitly in the class and throws this error.

Reviewers: NorthIsUp, branden

Reviewed By: branden

Subscribers: melsa

Differential Revision: http://phabricator.local.disqus.net/D18448