Report command failure properly in GenericXUnitTestEngine
Summary:
Currently, if a test command exits with a non-zero code, `GenericXunitTestEngine` assumes the command completely failed
and quits by throwing a hard to understand exception. This patch records this failure as `BROKEN: Unit Test Script` and tries to
make the output more human-friendly.

NOTE: We only output STDERR which is always empty when using Grunt because STDOUT contains the XML output which makes things
very hard to read and trace.

Test Plan:
Apply the patch, go to rNEXT, remove the `-f` flag from `.arcconfig`'s unit test command, introduce a syntax error or
test failure and inspect the result.

Reviewers: Andrew

Reviewed By: Andrew

Differential Revision: http://phabricator.local.disqus.net/D19100